requirements.txt
paddlepaddle
paddleocr
streamlit
Pillow
numpy
